Abstract getReferrer
